+++ Source/ASDisplayNode+Yoga.mm
@@ -298,7 +298,7 @@
 
   // Reset accessible elements, since layout may have changed.
   ASPerformBlockOnMainThread(^{
-    [(_ASDisplayView *)self.view setAccessibleElements:nil];
+    [(_ASDisplayView *)self.view setAccessibilityElements:nil];
   });
 
   ASDisplayNodePerformBlockOnEveryYogaChild(self, ^(ASDisplayNode * _Nonnull node) {